[M15] scheme: r7rs_core v0 corpus green on all 8 runners + the grid
Six cases — arith, tail-recursive factorial, mutual even/odd, and
three call/cc escapes (loop exit, threshold search, nested inner/outer)
— byte-identical across interp, jit×{arena,rc}, the AOT grid
(6/6 × x86_64/aarch64/riscv64/wasm32 × 2 + s390x canary), and the real
chibi-scheme 0.9.1 oracle. diff[…,scheme,…]: 77 cases, 0 divergent.

The escape cases are the milestone's point: call/cc lowers to
frk_ctl.prompt over an escape-closure, (k v) to frk_ctl.abort, and the
frontend-explicit guards make native result-passing propagate exactly
as the interpreter's real unwind — escape_nested's 105 (inner escape
leaves the outer +100) and 8 (inner escape to the outer prompt skips
it) match chibi on the nose.

Fix found by the grid: frk_rt_scm_display_bool took u8, but the
lowering passes the extended i1 as i64 — wasm's exact-import-signature
rule trapped it (the LANDSCAPE-pinned lesson). i64 in both twins + the
JIT capture now.

+++ crates/frk-harness/src/runner.rs
@@ -635,7 +635,7 @@ extern "C" fn capture_scm_display_num(value: f64) {
     CAPTURE.with(|buffer| buffer.borrow_mut().push_str(&frk_rt::format_lua_num(value)));
 }
 
-extern "C" fn capture_scm_display_bool(value: u8) {
+extern "C" fn capture_scm_display_bool(value: i64) {
     CAPTURE.with(|buffer| buffer.borrow_mut().push_str(if value != 0 { "#t" } else { "#f" }));
 }
